pyproject.toml
[project]
name = "quack-diff"
version = "0.1.0"
description = "The zero-dependency regression testing tool for modern data warehouses. DuckDB-powered data diffing across Snowflake, Postgres, and more."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Matteo Renoldi", email = "matteo@renoldi.org" }
]
requires-python = ">=3.11"
keywords = ["data", "diff", "duckdb", "snowflake", "postgres", "regression-testing", "data-quality"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "duckdb>=1.0.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
quack-diff = "quack_diff.cli.main:app"

[project.urls]
Homepage = "https://github.com/matteorenoldi/quack-diff"
Repository = "https://github.com/matteorenoldi/quack-diff"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
